# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: MC School Studio API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects/{projectId}/import/parse:
    post:
      operationId: parseImportFile
      tags: [import]
      summary: Upload and parse an Excel or CSV file, returns sheet/column preview
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/ImportFileForm"
      responses:
        "200":
          description: Parse result with sheets and sample rows
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ParseResult"
        "400":
          description: Invalid file
        "401":
          description: Unauthorized
  /projects/{projectId}/import/confirm:
    post:
      operationId: confirmImport
      tags: [import]
      summary: Confirm a project-scoped roster reconciliation
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImportConfirmation"
      responses:
        "200":
          description: Import result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportResult"
        "400":
          description: Validation error
        "401":
          description: Unauthorized
...
components:
  schemas:
...
    ParseResult:
      type: object
      required: [sheets]
      properties:
        sheets:
          type: array
          items:
            $ref: "#/components/schemas/SheetPreview"
    SheetPreview:
      type: object
      required: [name, headers, rows]
      properties:
        name:
          type: string
        headers:
          type: array
          items:
            type: string
        rows:
          type: array
          items:
            type: array
            items:
              type: string
    ImportConfirmation:
      type: object
      required: [sheets]
      properties:
        sheets:
          type: array
          items:
            $ref: "#/components/schemas/SheetMapping"
    SheetMapping:
      type: object
      required: [sheetName, className, firstNameColumn, lastNameColumn, rows]
      properties:
        sheetName:
          type: string
        className:
          type: string
        firstNameColumn:
          type: string
        lastNameColumn:
          type: string
        studentIdColumn:
          type: ["string", "null"]
        emailColumn:
          type: ["string", "null"]
        phoneColumn:
          type: ["string", "null"]
        secondaryEmailColumn:
          type: ["string", "null"]
        jobTitleColumn:
          type: ["string", "null"]
        officeLocationColumn:
          type: ["string", "null"]
        photoSessionColumn:
          type: ["string", "null"]
        captureNotesColumn:
          type: ["string", "null"]
        rows:
          type: array
          items:
            type: array
            items:
              type: string
        headers:
          type: array
          items:
            type: string
    ImportResult:
      type: object
      required: [classesCreated, studentsCreated, studentsUpdated, studentsMoved, studentsSkipped, conflicts]
      properties:
        classesCreated:
          type: integer
          description: Classes created; existing classes are reused case-insensitively.
        studentsCreated:
          type: integer
          description: Students or employees created.
        studentsUpdated:
          type: integer
          description: Existing students or employees whose roster details changed.
        studentsMoved:
          type: integer
          description: Existing students or employees moved to another class or department.
        studentsSkipped:
          type: integer
          description: Unchanged rows skipped because they were already reconciled.
        conflicts:
          type: integer
          description: Ambiguous rows that were not merged.
    ImportFileForm:
      type: object
      required: [file]
      properties:
        file:
          type: string
          format: binary
          description: Excel (.xlsx) or CSV (.csv) file
        csvClassName:
          type: string
          description: Required only for CSV files — the class name to assign
...